+++ packages/at_chops/example/at_chops_example.dart
@@ -8,10 +8,10 @@ import 'package:encrypt/encrypt.dart';
 
 /// Usage:
 /// Using new key pairs created at runtime
-/// dart run at_chops_example.dart
+/// `dart at_chops_example.dart`
 /// or
 /// Using key pairs from atKeys file
-/// dart run at_chops_example.dart <path_to_atkeys_file>
+/// `dart at_chops_example.dart <path_to_atkeys_file>`
 void main(List<String> args) async {
   AtChops atChops;
   if (args.isNotEmpty) {