+++ packages/at_server_status/test/at_server_status_test.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:at_server_status/at_server_status.dart';
-import 'package:at_server_status/at_status_impl.dart';
 import 'package:test/test.dart';
 import 'package:uuid/uuid.dart';