+++ packages/at_auth/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.1.0
+- fix: potential bug handling atSigns which end in `data` e.g. `@foo_data`
+
 ## 2.0.10
 - fix: Replace legacy IVs with random IVs for encrypting "defaultEncryptionPrivateKey" and "selfEncryptionKey" in APKAM flow
 ## 2.0.9